cli/zip: split the code in smaller functions and files
.. to ease review of subsequent changes.

No functional change in this commit, other than moving
the `doctor` call to below the per-node fetches.

Release note: None
